COM-1522: rename style
